refactor: update dependencies to 1.18.2 versions
