+++ build.gradle
@@ -59,11 +59,6 @@ subprojects {
                         curseforge = "roughly-enough-items"
                         modrinth = "rei"
                     }
-
-                    optional {
-                        curseforge = "jei"
-                        modrinth = "jei"
-                    }
                 }
 
                 var curseId = System.getenv("CURSEFORGE_ID")
