+++ common/src/main/java/net/impleri/fluidskills/integrations/kubejs/FluidSkillsPlugin.java
@@ -1,17 +1,13 @@
 package net.impleri.fluidskills.integrations.kubejs;
 
 import dev.latvian.mods.kubejs.KubeJSPlugin;
+import dev.latvian.mods.kubejs.script.ScriptType;
 import net.impleri.fluidskills.integrations.kubejs.events.EventsBinding;
 import net.impleri.fluidskills.integrations.kubejs.events.RestrictionsRegistrationEventJS;
 import net.minecraft.server.MinecraftServer;
 
 public class FluidSkillsPlugin extends KubeJSPlugin {
-    @Override
-    public void registerEvents() {
-        EventsBinding.GROUP.register();
-    }
-
     public static void onStartup(MinecraftServer minecraftServer) {
-        EventsBinding.RESTRICTIONS.post(new RestrictionsRegistrationEventJS(minecraftServer));
+        new RestrictionsRegistrationEventJS(minecraftServer).post(ScriptType.SERVER, EventsBinding.RESTRICTIONS);
     }
 }
