--- fabric/src/main/java/net/impleri/fluidskills/fabric/FluidSkillsFabric.java
@@ -1,11 +1,0 @@
-package net.impleri.fluidskills.fabric;
-
-import net.fabricmc.api.ModInitializer;
-import net.impleri.fluidskills.FluidSkills;
-
-public class FluidSkillsFabric implements ModInitializer {
-    @Override
-    public void onInitialize() {
-        FluidSkills.init();
-    }
-}
