--- fabric/src/main/java/net/impleri/fluidskills/integrations/jei/fabric/FluidSkillsJeiPluginFabric.java
@@ -1,11 +1,0 @@
-package net.impleri.fluidskills.integrations.jei.fabric;
-
-import me.shedaniel.rei.plugincompatibilities.api.REIPluginCompatIgnore;
-import mezz.jei.api.JeiPlugin;
-import mezz.jei.api.fabric.ingredients.fluids.IJeiFluidIngredient;
-import net.impleri.fluidskills.integrations.jei.FluidSkillsJeiPlugin;
-
-@REIPluginCompatIgnore()
-@JeiPlugin()
-public class FluidSkillsJeiPluginFabric extends FluidSkillsJeiPlugin<IJeiFluidIngredient> {
-}
