+++ forge/build.gradle
@@ -35,9 +35,8 @@ dependencies {
     compileOnly "me.shedaniel:RoughlyEnoughItems-default-plugin-forge:${rootProject.rei_version}"
     modImplementation "me.shedaniel:RoughlyEnoughItems-forge:${rootProject.rei_version}"
 
-    compileOnly "mezz.jei:jei-${rootProject.minecraft_version}-common-api:${rootProject.jei_version}"
-    compileOnly "mezz.jei:jei-${rootProject.minecraft_version}-forge-api:${rootProject.jei_version}"
-    modImplementation "mezz.jei:jei-${rootProject.minecraft_version}-forge:${rootProject.jei_version}"
+    compileOnly "mezz.jei:jei-${rootProject.minecraft_version}:${rootProject.jei_version}:api"
+    modImplementation "mezz.jei:jei-${rootProject.minecraft_version}:${rootProject.jei_version}"
 
     common(project(path: ":common", configuration: "namedElements")) { transitive false }
     shadowCommon(project(path: ":common", configuration: "transformProductionForge")) { transitive = false }
@@ -86,6 +85,17 @@ publishing {
         mavenForge(MavenPublication) {
             artifactId = rootProject.archives_base_name + "-" + project.name
             from components.java
+        }
+    }
+}
+
+unifiedPublishing {
+    project {
+        relations {
+            optional {
+                curseforge = "jei"
+                modrinth = "jei"
+            }
         }
     }
 }
