+++ forge/src/main/resources/fluidskills.mixins.json
@@ -3,9 +3,7 @@
   "minVersion": "0.8",
   "package": "net.impleri.fluidskills.mixins.forge",
   "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "MixinEntity"
-  ],
+  "mixins": [],
   "client": [],
   "server": [],
   "injectors": {
