+++ gradle.properties
@@ -1,21 +1,21 @@
 org.gradle.jvmargs=-Xmx4096m
 maven_group=net.impleri
-minecraft_version=1.19.2
+minecraft_version=1.18.2
 # Architectury-related config
 enabled_platforms=fabric,forge
-archives_base_name=fluid-skills-1.19.2
-architectury_version=6.4.62
+archives_base_name=fluid-skills-1.18.2
+architectury_version=4.11.89
 # Fabric config
-fabric_loader_version=0.14.13
-fabric_api_version=0.72.0+1.19.2
+fabric_loader_version=0.14.17
+fabric_api_version=0.75.1+1.18.2
 # Forge config
-forge_version=1.19.2-43.2.0
+forge_version=1.18.2-40.2.0
 # KubeJS config
-kubejs_version=1902.6.0-build.142
-rhino_version=1902.2.2-build.264
+kubejs_version=1802.5.5-build.569
+rhino_version=1802.2.1-build.255
 # PlayerSkills dependency
-playerskills_version=1.7.2
+playerskills_version=1.0.2
 # REI Integration
-rei_version=9.1.587
+rei_version=8.3.590
 # JEI Integration
-jei_version=11.4.0.286
+jei_version=9.7.1.255
